+++ CHANGELOG.md
@@ -1,4 +1,6 @@
-## (WIP) v0.15.1
+## v0.15.1
+
+- Trigger the related `Record` model realtime subscription events on [custom model struct](https://pocketbase.io/docs/custom-models/) save ([#2325](https://github.com/pocketbase/pocketbase/discussions/2325)).
 
 - Fixed `Ctrl + S` in the `editor` field not propagating the quick save shortcut to the parent form.
 
@@ -8,7 +10,7 @@
 
 - Reduced the record form vertical layout shifts and slightly improved the rendering speed when loading multiple `relation` fields.
 
-- Trigger the related `Record` model realtime subscription events on [custom model struct](https://pocketbase.io/docs/custom-models/) save ([#2325](https://github.com/pocketbase/pocketbase/discussions/2325)).
+- Enabled Admin UI assets cache.
 
 
 ## v0.15.0